async stuff
need to move the calls to async to be staggered / dependent on already running tasks.

think we are spawning too many tasks at once when loading textures off the back of a model. So when we "LoadAsset" instead we will push the handle into a queue and update checks to ensure that it will still return loading, even though the task hasnt started yet.